fix: `GROUP_CONCAT` handling of empty separator (issue) (#2474)
`GROUP_CONCAT` was handling an empty separator (i.e. `""`) incorrectly,
it would handle it as if the separator were not set, so essentially it was
treated as a single space (i.e. `" "`).

This change fixes it so that an empty separator with `GROUP_CONCAT`
results in a value with nothing between concatenated values.


Fixes <https://github.com/RDFLib/rdflib/issues/2473>

---------

Co-authored-by: WhiteGobo <richardfechner@posteo.net>