test: add skips to accommodate a DBpedia outage
DBpedia is failing right now. This change adds narrow skips to the tests
that use it which only trigger for connection related problems, so the
test suite can pass even when DBpedia is down.

Test that use external services should be avoided if possible and should
be replaced with the HTTP mocking facilities of the test framework.