build: replace Flake8, FlakeHeaven and isort with ruff
FlakeHeaven only supports Flake8 versions below 5.x, and flake8 itself
is still fairly limited in its configurability.

Ruff implements most of the features of isort, Flake8 and FlakeHeaven.
It does not provide a baseline feature like FlakeHeaven, but it does
have a feature for adding `# noqa:` directives to the code base, which
was used after file specific ignores were configured for some errors.

Ruff also has a fix feature which can be used to fix some errors, so
this change also adds Ruff to the pre-commit config so that it can be
triggered with `pre-commit.ci autofix`.