fix: SPARQL `LOAD ... INTO GRAPH` handling (#2554)
`LOAD ... INTO GRAPH` stopped working correctly after the change to
handling of the `publicID` `Graph.parse` parameter in RDFLib 7.0.0
(<https://github.com/RDFLib/rdflib/pull/2406>).

This is because `LOAD` evaluation relied on `publicID` to select the
graph name. So after <https://github.com/RDFLib/rdflib/pull/2406> data
would be loaded into the default graph even if a named graph is
specified.

This change adds tests for `LOAD ... INTO GRAPH` and fixes the load
evaluation.

A consequence of this change is also that relative IRI lookup for graphs
loaded with `LOAD ... INTO GRAPH` is now relative to the source document
URI instead of the base URI of the graph being loaded into, which is
more correct.