style: eliminate unused `noqa` statements (#2566)
This change enables RUF100 ("Unused noqa directive") and removes all the
`noqa` directives that are no longer needed.

This is done with `poetry run ruff check --fix .`, which also removes
`noqa` directives that are not needed because of ruff config, like E402.

This change also removes file level ignores for some python files in
favour of `noqa` directives. This is to avoid removing `noqa` directives
that will just be added back later.