Add changelog to sphinx docs (#2617)
* docs: include changelog in docs

* docs: update changelog headings for sphinx rendering

With the previous commit adding the changelog to the docs, there were
issues with how the MyST parser and Sphinx were handling this. Because
there were multiple 'h1' level heading, this inserted a line in the
toctree for every h1 heading, i.e. every release, rather than a single
link to the changelog page.

The '===' were also rendering literally in the Sphinx output, so I
changed these.

Levels of headings were also changed to consistently move downwards so
releases are the only h2 headings, and headings that relate to a
specific part of a release are h3.