fix: Prevent Collection from add 'rdf:nil rdf:rest rdf:nil.' triples
Creating 'Collection(graph, RDF.nil)' used to add a 'rdf:nil rdf:rest rdf:nil.'
triple to the graph, which turned the empty list into an infinite list.

This was a side-effect of unconditionally appending 'seq' to the collection,
which is not possible when the collection is 'nil' without ill side-effects.

This commit makes the append conditional; and also checks the collection
is non-empty in case the user explicitly tries to add item to the empty list.