Downgrade log message about  plugin without override argument to debug from warning (#3063)
Original issue

- https://github.com/RDFLib/rdflib/issues/1880

was worked around by providing a fall back call without "override" in

- https://github.com/RDFLib/rdflib/pull/2018

As a result any user with such a store (in our case just <class
'rdflib.plugins.memory.IOMemory'>) is flooded with such warnings pointing to
that issue #1880 and for all means and purposes AFAIK just need to ignore
them since there is nothing user can do to address it.

That raises the question on why is it a Warning at a user level and not some
debug message for developers/troubleshooting etc?  Hence I am lowering it to
debug level to bring peace of mind to users of the library.

Co-authored-by: Nicholas Car <nick@kurrawong.net>