fix: remove Literal.toPython date conversion for gYear/gYearMonth (#3115)
* fix: remove Literal.toPython casting for gYear and gYearMonth

Issue #3078 reports, that rdflib.Literal.toPython casting of xsd:gYear and
xsd:gYearMonth to datetime objects is not possible, as there is no
appropriate Python equivalence for those types.

The current implementation casts xsd:gYear and xsd:gYearMonth to
datetime objects assuming January 1st for xsd:gYear and the 1st day
of the given month for xsd:gYearMonth. This is plain wrong.

The change removes casting to datetime objects in
rdflib.Literal.toPython for xsd:gYear and xsd:gYearMonth.

Closes #3078 .

* test: adapt rdflib.Literal tests to gYear/gYearMonth toPython change

---------

Co-authored-by: Nicholas Car <nick@kurrawong.net>